==> notebooks/ml/01_first_look.ipynb <==
# Deep Neural Network Model

**Resources**
- Baseline code from [tensorflow github](https://github.com/tensorflow/recommenders/blob/main/docs/examples/context_features.ipynb)
    - ❓ Are we sure that we train each entry with the specific movie and not the listo of movies watched from the user

!pip install -q tensorflow-recommenders

import os
import tempfile

import numpy as np
import tensorflow as tf
import tensorflow_recommenders as tfrs

## TODO: remove this cells
- Explore the example dataset

# TODO remove this cell
!pip install -q --upgrade tensorflow-datasets
import tensorflow_datasets as tfds

ratings = tfds.load("movielens/100k-ratings", split="train")
movies = tfds.load("movielens/100k-movies", split="train")

ratings = ratings.map(lambda x: {
    "movie_title": x["movie_title"],
    "user_id": x["user_id"],
    "timestamp": x["timestamp"],
})
movies = movies.map(lambda x: x["movie_title"]) # Target?



timestamps = np.concatenate(list(ratings.map(lambda x: x["timestamp"]).batch(100)))

max_timestamp = timestamps.max()
min_timestamp = timestamps.min()

timestamp_buckets = np.linspace(
    min_timestamp, max_timestamp, num=1000,
)

unique_movie_titles = np.unique(np.concatenate(list(movies.batch(1000))))
unique_user_ids = np.unique(np.concatenate(list(ratings.batch(1_000).map(
    lambda x: x["user_id"]))))


type(movies), type(ratings)

**TODO**
- ❓ why we have both movies and ratings? we can extract movies directly from ratings, or not?

type(unique_user_ids), unique_user_ids[0]

type(unique_movie_titles), unique_movie_titles[0]

movies.as_numpy_iterator().next()

ratings.as_numpy_iterator().next()

## Load data

**TODO**
- Store dataset info like all the product id with the dataset

df_train["User_ID"] = df_train["User_ID"].astype(str)
df_train["Gender"] = df_train["Gender"].astype(str)
df_train["Product_ID"] = df_train["Product_ID"].astype(str)

import pandas as pd
from os.path import join

dataset_parsed_path = "/home/jupyter/mlspec-blackfriday/dataset/parsed/202104130952/"
train_path = join(dataset_parsed_path, "train.csv")
test_path = join(dataset_parsed_path, "eval.csv")

df_train=pd.read_csv(train_path)
df_test=pd.read_csv(test_path)

print(f"df_train shape:{df_train.shape}")
print(f"df_test shape:{df_test.shape}")

df_train.head(3)

**TODO**
- TODO calculate and store those info with all the data, with the data

unique_gender = df_train["Gender"].unique()
type(unique_gender), len(unique_gender), unique_gender[0]

unique_user_id = df_train["User_ID"].unique()
type(unique_user_id), len(unique_user_id), unique_user_id[0]

unique_product_id = df_train["Product_ID"].unique()
type(unique_product_id), len(unique_product_id), unique_product_id[0]

## Model definition

### User model

- Require:
    - All the values of features that need `StringLookup` in `numpy.ndarray` format
    
**TODO** 
- extract embeddings dimension as parameter

class UserModel(tf.keras.Model):
  
    def __init__(self):
        super().__init__()

        self.user_embedding = tf.keras.Sequential([
            tf.keras.layers.experimental.preprocessing.StringLookup(
                vocabulary=unique_user_id, mask_token=None),
            tf.keras.layers.Embedding(len(unique_user_id) + 1, 32),
        ])

        self.gender_embedding = tf.keras.Sequential([
            tf.keras.layers.experimental.preprocessing.StringLookup(
                    vocabulary=unique_gender, mask_token=None),
            tf.keras.layers.Embedding(len(unique_gender) + 1, 32),
        ])


    def call(self, inputs):
        return tf.concat([
            self.user_embedding(inputs["user_id"]),
            self.gender_embedding(inputs["Gender"]),
        ], axis=1)

## Product model
- Simple start: use only product id

**TODO**
- Find out what `max_tokens` we should use / or if we can remove it
- Parametrize the embeddings dimension

class ProductModel(tf.keras.Model):
  
    def __init__(self):
        super().__init__()

        max_tokens = 10_000

        self.product_embedding = tf.keras.Sequential([
          tf.keras.layers.experimental.preprocessing.StringLookup(
              vocabulary=unique_product_id, mask_token=None),
          tf.keras.layers.Embedding(len(unique_product_id) + 1, 32)
        ])


    def call(self, products_id):
        return tf.concat([
            self.product_embedding(products_id)
        ], axis=1)

## Combined model

**Info**
-  `FactorizedTopK` take the list of candidates for build the metric [link](https://www.tensorflow.org/recommenders/api_docs/python/tfrs/metrics/FactorizedTopK)

**TODO**
- Parametrize the embedding dimension
- ❓ Explore the model: we have created two layers of 32 neurons?
- ❓ Ensure `FactorizedTopK` is well understood

# TODO refactory and move this code
tf_products = tf.data.Dataset.from_tensor_slices(df_train["Product_ID"].astype(str).values)

class BlackFridayModel(tfrs.models.Model):

    def __init__(self):
        super().__init__()
        self.query_model = tf.keras.Sequential([
          UserModel(),
          tf.keras.layers.Dense(32)
        ])
        self.candidate_model = tf.keras.Sequential([
          ProductModel(),
          tf.keras.layers.Dense(32)
        ])
        self.task = tfrs.tasks.Retrieval(
            metrics=tfrs.metrics.FactorizedTopK(
                candidates=tf_products.batch(128).map(self.candidate_model), # dataset of candidate embeddings from which candidates should be retrieved (embedded)
            ),
        )

    def compute_loss(self, features, training=False):
        query_embeddings = self.query_model({
            "User_ID": features["User_ID"],
            "Gender": features["Gender"],
        })
        product_embeddings = self.candidate_model(features["movie_title"])

        return self.task(query_embeddings, product_embeddings)

## Prepare the data

df_train_parsed = df_train[["User_ID", "Gender", "Product_ID"]]
df_train_parsed.head()

df_train["User_ID"] = df_train["User_ID"].astype(str)
df_train["Gender"] = df_train["Gender"].astype(str)
df_train["Product_ID"] = df_train["Product_ID"].astype(str)

df_train_parsed = df_train[["User_ID", "Gender", "Product_ID"]]
df_train_parsed = df_train_parsed.reset_index(drop=True)
df_train_parsed.dtypes

# TODO refactory and move this code
tf_dataset = tf.data.Dataset.from_tensor_slices(df_train_parsed.to_numpy())
tf_dataset

tf.random.set_seed(42)
shuffled = tf_dataset.shuffle(100_000, seed=42, reshuffle_each_iteration=False)

train = shuffled.take(80_000)
test = shuffled.skip(80_000).take(20_000)

cached_train = train.shuffle(100_000).batch(2048)
cached_test = test.batch(4096).cache()

model = BlackFridayModel()

model = BlackFridayModel()
model.compile(optimizer=tf.keras.optimizers.Adagrad(0.1))

model.fit(cached_train, epochs=3)

train_accuracy = model.evaluate(
    cached_train, return_dict=True)["factorized_top_k/top_100_categorical_accuracy"]
test_accuracy = model.evaluate(
    cached_test, return_dict=True)["factorized_top_k/top_100_categorical_accuracy"]
    
print(f"Top-100 accuracy (train): {train_accuracy:.2f}.")
print(f"Top-100 accuracy (test): {test_accuracy:.2f}.")
