==> notebooks/ml/06_trained_models_loader.ipynb <==
# Load trained models

> use this notebook to check the correctness of tained models

job_dir = "gs://mlteam-ml-specialization-2021-blackfriday/aiplatform_jobs/bf_20210426_145558/"

from os.path import join

scann_path = join(job_dir, "Scann")

## Load Scann

import tensorflow as tf
import tensorflow_recommenders as tfrs # requirea to load SCANN
scann_loaded = tf.keras.models.load_model(scann_path)

scann_example = {  # used to check scann consistency
    "Gender": tf.constant(["M"]),
    "Age": tf.constant(["26-35"]),
    "Occupation": tf.constant(["0"]),
    "City_Category": tf.constant(["B"]),
    "Stay_In_Current_City_Years": tf.constant(["4+"]),
    "Marital_Status": tf.constant(["0"]),
}


scann_results = scann_loaded(scann_example)  # prediction required to store the model

scann_results

### Check if the prediction match:

!gsutil cat $"{job_dir}"params.json
