==> notebooks/ml/08_split_eval_test.ipynb <==
from sklearn.model_selection import StratifiedShuffleSplit
import pandas as pd

test_data_path="gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test.csv"
label_column="Product_ID"

df=pd.read_csv(test_data_path)
df

# removing products with only one instance

label_numerosities=df[label_column].value_counts()
multiple_products = label_numerosities[label_numerosities>1].index
clean_df = df[df[label_column].isin(multiple_products)].reset_index(drop=True)

feature_cols = [c for c in clean_df.columns if c != label_column]
label_cols   = [label_column]
x, y = clean_df[feature_cols].values, clean_df[label_cols].values
for eval_index, test_index in StratifiedShuffleSplit(n_splits=1, test_size=0.5, random_state=42).split(x,y):
    print((eval_index.shape, test_index.shape))
    df_eval_clean=clean_df.iloc[eval_index]
    df_test_clean=clean_df.iloc[test_index]

### sanity checks

pd.DataFrame({
    "eval":df_eval_clean[label_column].value_counts(),
    "test":df_test_clean[label_column].value_counts()
})

(df_test_clean[label_column].value_counts() - df_eval_clean[label_column].value_counts()).value_counts()

### handling single products

single_label_df=df[~df[label_column].isin(multiple_products)].reset_index(drop=True)
single_label_eval_df = single_label_df.iloc[int(single_label_df.shape[0]/2):]
single_label_test_df = single_label_df.iloc[:int(single_label_df.shape[0]/2)] 

df_eval = pd.concat([df_eval_clean, single_label_eval_df])
df_test = pd.concat([df_test_clean, single_label_test_df])

### saving results

df_eval.to_csv("gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test/evalset.csv")
df_eval.to_csv("gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test/testset.csv")
