==> notebooks/ml/09_deployment.ipynb <==
%%writefile prediction_input.json
{"Gender": ["M"],"Age": ["26-35"],"Occupation": ["0"],"City_Category": ["B"],"Stay_In_Current_City_Years": ["4+"],"Marital_Status": ["0"]}


%%bash

MODEL_DIR="gs://mlteam-ml-specialization-2021-blackfriday/aiplatform_jobs/004_hptuning_ed_allfeatures/3/Scann"

gcloud ai-platform local predict --model-dir ${MODEL_DIR}/ \
  --json-instances prediction_input.json \
  --framework tensorflow \
  --signature-name='serving_default'

!gcloud ai-platform local predict --help

import tensorflow as tf
import tensorflow_recommenders as tfrs

model_dir="gs://mlteam-ml-specialization-2021-blackfriday/aiplatform_jobs/004_hptuning_ed_allfeatures/3/Scann"

model=tf.keras.models.load_model(model_dir)
model

model.signatures

print("""ERROR: (gcloud.ai-platform.local.predict) b'2021-05-31 12:57:07.471041: I tensorflow/stream_executor/platform/default/dso_loader.cc:49] Successfully opened dynamic library libcudart.so.11.0\nWARNING:tensorflow:From /opt/conda/lib/python3.7/site-packages/tensorflow/python/compat/v2_compat.py:96: disable_resource_variables (from tensorflow.python.ops.variable_scope) is deprecated and will be removed in a future version.\nInstructions for updating:\nnon-resource variables are not supported in the long term\n2021-05-31 12:57:09.706238: I tensorflow/compiler/jit/xla_cpu_device.cc:41] Not creating XLA devices, tf_xla_enable_xla_devices not set\n2021-05-31 12:57:09.706591: W tensorflow/stream_executor/platform/default/dso_loader.cc:60] Could not load dynamic library \'libcuda.so.1\'; dlerror: libcuda.so.1: cannot open shared object file: No such file or directory; LD_LIBRARY_PATH: /usr/local/cuda/lib64:/usr/local/nccl2/lib:/usr/local/cuda/extras/CUPTI/lib64\n2021-05-31 12:57:09.706668: W tensorflow/stream_executor/cuda/cuda_driver.cc:326] failed call to cuInit: UNKNOWN ERROR (303)\n2021-05-31 12:57:09.706704: I tensorflow/stream_executor/cuda/cuda_diagnostics.cc:156] kernel driver does not appear to be running on this host (blackfriday-tensorflow-2-3-20210407-151403): /proc/driver/nvidia/version does not exist\n2021-05-31 12:57:09.707547: I tensorflow/compiler/jit/xla_gpu_device.cc:99] Not creating XLA devices, tf_xla_enable_xla_devices not set\nWARNING:tensorflow:From /usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/frameworks/tf_prediction_lib.py:236: load (from tensorflow.python.saved_model.loader_impl) is deprecated and will be removed in a future version.\nInstructions for updating:\nThis function will only be available through the v1 compatibility library as tf.compat.v1.saved_model.loader.load or tf.compat.v1.saved_model.load. There will be a new function for importing SavedModels in Tensorflow 2.0.\nWARNING:tensorflow:From /usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/frameworks/tf_prediction_lib.py:236: load (from tensorflow.python.saved_model.loader_impl) is deprecated and will be removed in a future version.\nInstructions for updating:\nThis function will only be available through the v1 compatibility library as tf.compat.v1.saved_model.loader.load or tf.compat.v1.saved_model.load. There will be a new function for importing SavedModels in Tensorflow 2.0.\nERROR:root:Failed to load the model due to bad model data. tags: [\'serve\']\nTraceback (most recent call last):\n  File "/usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/frameworks/tf_prediction_lib.py", line 236, in load_tf_model\n    session, tags=list(tags), export_dir=model_path)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/util/deprecation.py", line 340, in new_func\n    return func(*args, **kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/saved_model/loader_impl.py", line 300, in load\n    return loader.load(sess, tags, import_scope, **saver_kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/saved_model/loader_impl.py", line 453, in load\n    **saver_kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/saved_model/loader_impl.py", line 383, in load_graph\n    meta_graph_def, import_scope=import_scope, **saver_kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/training/saver.py", line 1485, in _import_meta_graph_with_return_elements\n    **kwargs))\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/framework/meta_graph.py", line 804, in import_scoped_meta_graph_with_return_elements\n    return_elements=return_elements)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/util/deprecation.py", line 538, in new_func\n    return func(*args, **kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/framework/importer.py", line 405, in import_graph_def\n    producer_op_list=producer_op_list)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/framework/importer.py", line 497, in _import_graph_def_internal\n    graph._c_graph, serialized, options)  # pylint: disable=protected-access\ntensorflow.python.framework.errors_impl.NotFoundError: Op type not registered \'Scann>ScannSearchBatched\' in binary running on blackfriday-tensorflow-2-3-20210407-151403. Make sure the Op and Kernel are registered in the binary running in this process. Note that if you are loading a saved graph which used ops from tf.contrib, accessing (e.g.) `tf.contrib.resampler` should be done before importing the graph, as contrib ops are lazily registered when the module is first accessed.\nTraceback (most recent call last):\n  File "/usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/frameworks/tf_prediction_lib.py", line 236, in load_tf_model\n    session, tags=list(tags), export_dir=model_path)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/util/deprecation.py", line 340, in new_func\n    return func(*args, **kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/saved_model/loader_impl.py", line 300, in load\n    return loader.load(sess, tags, import_scope, **saver_kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/saved_model/loader_impl.py", line 453, in load\n    **saver_kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/saved_model/loader_impl.py", line 383, in load_graph\n    meta_graph_def, import_scope=import_scope, **saver_kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/training/saver.py", line 1485, in _import_meta_graph_with_return_elements\n    **kwargs))\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/framework/meta_graph.py", line 804, in import_scoped_meta_graph_with_return_elements\n    return_elements=return_elements)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/util/deprecation.py", line 538, in new_func\n    return func(*args, **kwargs)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/framework/importer.py", line 405, in import_graph_def\n    producer_op_list=producer_op_list)\n  File "/opt/conda/lib/python3.7/site-packages/tensorflow/python/framework/importer.py", line 497, in _import_graph_def_internal\n    graph._c_graph, serialized, options)  # pylint: disable=protected-access\ntensorflow.python.framework.errors_impl.NotFoundError: Op type not registered \'Scann>ScannSearchBatched\' in binary running on blackfriday-tensorflow-2-3-20210407-151403. Make sure the Op and Kernel are registered in the binary running in this process. Note that if you are loading a saved graph which used ops from tf.contrib, accessing (e.g.) `tf.contrib.resampler` should be done before importing the graph, as contrib ops are lazily registered when the module is first accessed.\n\nDuring handling of the above exception, another exception occurred:\n\nTraceback (most recent call last):\n  File "/usr/lib/google-cloud-sdk/lib/googlecloudsdk/command_lib/ml_engine/local_predict.py", line 184, in <module>\n    main()\n  File "/usr/lib/google-cloud-sdk/lib/googlecloudsdk/command_lib/ml_engine/local_predict.py", line 179, in main\n    signature_name=args.signature_name)\n  File "/usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/prediction_lib.py", line 98, in local_predict\n    client = create_client(framework, model_dir, **kwargs)\n  File "/usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/prediction_lib.py", line 91, in create_client\n    return create_client_fn(model_path, **kwargs)\n  File "/usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/frameworks/tf_prediction_lib.py", line 550, in create_tf_session_client\n    return SessionClient(*load_tf_model(model_dir, tags, config))\n  File "/usr/lib/google-cloud-sdk/lib/third_party/ml_sdk/cloud/ml/prediction/frameworks/tf_prediction_lib.py", line 242, in load_tf_model\n    "%s\\n%s" % (msg, str(e)))\ncloud.ml.prediction.prediction_utils.PredictionError: Failed to load model: Failed to load the model due to bad model data. tags: [\'serve\']\nOp type not registered \'Scann>ScannSearchBatched\' in binary running on blackfriday-tensorflow-2-3-20210407-151403. Make sure the Op and Kernel are registered in the binary running in this process. Note that if you are loading a saved graph which used ops from tf.contrib, accessing (e.g.) `tf.contrib.resampler` should be done before importing the graph, as contrib ops are lazily registered when the module is first accessed. (Error code: 0)\n'""")

# new approach
Tutorials:  
https://github.com/google-research/google-research/tree/master/scann/tf_serving  
https://www.tensorflow.org/tfx/serving/docker#developing_with_docker  
https://cloud.google.com/ai-platform/prediction/docs/use-custom-container  

## (execute the following cells in terminals)

### making it work locally

%% bash
cd /tmp/Scann
gsutil -m cp -r gs://mlteam-ml-specialization-2021-blackfriday/aiplatform_jobs/010_hptuning_input_l1/1/Scann .
mv Scann 01
cd ..

!docker pull google/tf-serving-scann

%%bash
!docker run -t --rm -p 8501:8501 \
    -v "$(pwd)/Scann:/models/Scann" \
    -e MODEL_NAME=Scann \
    google/tf-serving-scann

curl -d '{"instances": [{"Gender":"M", "Age":"26-35", "Occupation":"0", "City_Category":"B", "Stay_In_Current_City_Years":"4+", "Marital_Status":"0"}]}' \
    -X POST http://localhost:8501/v1/models/Scann:predict

curl -X GET http://localhost:8501/v1/models/Scann

### creating a custom image

%%bash

docker run -d --name serving_base google/tf-serving-scann
docker cp Scann serving_base:/models/Scann
docker commit --change "ENV MODEL_NAME Scann" serving_base scann000
docker kill serving_base

docker run -t --rm -p 8501:8501 scann000

### pushing the image to Artifact Registry 

%%bash

gcloud auth configure-docker \
    europe-west1-docker.pkg.dev

%%bash
docker tag scann000 europe-west1-docker.pkg.dev/mlteam-ml-specialization-2021/scann-europe-west1/scann000
docker push europe-west1-docker.pkg.dev/mlteam-ml-specialization-2021/scann-europe-west1/scann000

### deploy the image to ai platform prediction

gcloud ai-platform models create scann --region=europe-west1  --enable-logging

gcloud beta ai-platform versions create scann000 \
  --region=europe-west1 \
  --model=scann \
  --machine-type=n1-standard-4 \
  --image=europe-west1-docker.pkg.dev/mlteam-ml-specialization-2021/scann-europe-west1/scann000 \
  --ports=8501 \
  --health-route=/v1/models/Scann \
  --predict-route=/v1/models/Scann:predict

curl -d '{"instances": [{"Gender":"M", "Age":"26-35", "Occupation":"0", "City_Category":"B", "Stay_In_Current_City_Years":"4+", "Marital_Status":"0"}]}' \
    -X POST https://ml.googleapis.com/v1/projects/mlteam-ml-specialization-2021/models/scann:predict

# the authentication problem
https://developers.google.com/identity/sign-in/web/devconsole-project  
https://stackoverflow.com/questions/44087267/google-api-machine-learning-can-i-use-an-api-key  
actual solution:  
https://stackoverflow.com/questions/57920885/exposing-api-endpoints-for-ai-application-deployed-on-google-ai-platform

%%bash
curl -d '{"instances": [{"Gender":"M", "Age":"26-35", "Occupation":"0", "City_Category":"B", "Stay_In_Current_City_Years":"4+", "Marital_Status":"0"}]}' \
    -X POST https://ml.googleapis.com/v1/projects/mlteam-ml-specialization-2021/models/scann:predict?access_token\="$(gcloud auth application-default print-access-token)"

%%bash
curl -d '{"instances": [{"Gender":"M", "Age":"26-35", "Occupation":"0", "City_Category":"B", "Stay_In_Current_City_Years":"4+", "Marital_Status":"0"}]}' \
    -X POST https://ml.googleapis.com/v1/projects/mlteam-ml-specialization-2021/models/Scann:predict?access_token\="$(gcloud auth application-default print-access-token)"

%%bash
curl -d '{"instances": [{"Gender":"M", "Age":"26-35", "Occupation":"0", "City_Category":"B", "Stay_In_Current_City_Years":"4+", "Marital_Status":"0"}]}' \
    -X POST https://europe-west1-ml.googleapis.com/v1/projects/mlteam-ml-specialization-2021/models/scann:predict?access_token\="$(gcloud auth application-default print-access-token)"
