==> notebooks/ml/10_product_categories.ipynb <==
from tqdm.notebook import tqdm
import pandas as pd
import numpy as np

train_source_path = "gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/train.csv"
eval_source_path  = "gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test/evalset.csv"
test_source_path  = "gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test/testset.csv"

train_dest_path  = "gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/preprocessed_category_train_.csv"
eval_dest_path   = "gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test/preprocessed_category_evalset.csv"
test_dest_path   = "gs://mlteam-ml-specialization-2021-blackfriday/dataset/parsed/202104130952/test/preprocessed_category_testset.csv"

## sanity checks

df = pd.read_csv(train_source_path)
df.isna().any()

df = pd.read_csv(eval_source_path)
df.isna().any()

df = pd.read_csv(test_source_path)
df.isna().any()

## preprocessing

def update_csv(source_path, dest_path):
    df = pd.read_csv(source_path)
    df = df.drop(columns=[x for x in df.columns if x.startswith("Unnamed")])
    df["Product_Category_1.2"]=df[["Product_Category_1","Product_Category_2"]].apply(list, axis=1).apply(
        lambda x: ".".join([str(int(y)) for y in x if not np.isnan(y)]))
    df["Product_Category_1.2.3"]=df[["Product_Category_1","Product_Category_2","Product_Category_3"]].apply(list, axis=1).apply(
        lambda x: ".".join([str(int(y)) for y in x if not np.isnan(y)]))
    df.to_csv(dest_path, index=False)

for source_path, dest_path in tqdm([
    (train_source_path, train_dest_path),
    (eval_source_path,  eval_dest_path),
    (test_source_path,  test_dest_path)
]):
    update_csv(source_path, dest_path)
print("done")
